+++ examples/restful-basic-authentication.go
@@ -2,8 +2,14 @@ package main
 
 import (
 	"github.com/emicklei/go-restful"
+	"io"
 	"net/http"
 )
+
+// This example shows how to create a (Route) Filter that performs Basic Authentication on the Http request.
+//
+// GET http://localhost:8080/secret
+// and use admin,admin for the credentials
 
 func main() {
 	ws := new(restful.WebService)
@@ -16,15 +22,14 @@ func basicAuthenticate(req *restful.Request, resp *restful.Response, chain *rest
 	encoded := req.Request.Header.Get("Authorization")
 	// usr/pwd = admin/admin
 	// real code does some decoding
-	if len(encoded) == 0 || "YWRzOmFzZA==" != encoded {
+	if len(encoded) == 0 || "Basic YWRtaW46YWRtaW4=" != encoded {
 		resp.AddHeader("WWW-Authenticate", "Basic realm=Protected Area")
-		resp.WriteHeader(401)
-		resp.Write([]byte("401: Not Authorized"))
+		resp.WriteErrorString(401, "401: Not Authorized")
 		return
 	}
 	chain.ProcessFilter(req, resp)
 }
 
 func secret(req *restful.Request, resp *restful.Response) {
-	resp.Write([]byte("42"))
+	io.WriteString(resp, "42")
 }