+++ examples/restful-user-resource.go
@@ -7,19 +7,34 @@ import (
 	"net/http"
 )
 
-// This example is functionally the same as the example in restful-user-resource.go
-// with the only difference that is served using its the restful.DefaultContainer
+// This example show a complete (GET,PUT,POST,DELETE) conventional example of
+// a REST Resource including documentation to be served by e.g. a Swagger UI
+// It is recommended to create a Resource struct (UserResource) that can encapsulate
+// an object that provide domain access (a DAO)
+// It has a Register method including the complete Route mapping to methods together
+// with all the appropriate documentation
+//
+// POST http://localhost:8080/users
+// <User><Id>1</Id><Name>Melissa Raspberry</Name></User>
+//
+// GET http://localhost:8080/users/1
+//
+// PUT http://localhost:8080/users/1
+// <User><Id>1</Id><Name>Melissa</Name></User>
+//
+// DELETE http://localhost:8080/users/1
+//
 
 type User struct {
 	Id, Name string
 }
 
-type UserService struct {
+type UserResource struct {
 	// normally one would use DAO (data access object)
 	users map[string]User
 }
 
-func (u UserService) Register() {
+func (u UserResource) Register(container *restful.Container) {
 	ws := new(restful.WebService)
 	ws.
 		Path("/users").
@@ -50,12 +65,12 @@ func (u UserService) Register() {
 		Doc("delete a user").
 		Param(ws.PathParameter("user-id", "identifier of the user").DataType("string")))
 
-	restful.Add(ws)
+	container.Add(ws)
 }
 
 // GET http://localhost:8080/users/1
 //
-func (u UserService) findUser(request *restful.Request, response *restful.Response) {
+func (u UserResource) findUser(request *restful.Request, response *restful.Response) {
 	id := request.PathParameter("user-id")
 	usr := u.users[id]
 	if len(usr.Id) == 0 {
@@ -68,7 +83,7 @@ func (u UserService) findUser(request *restful.Request, response *restful.Respon
 // POST http://localhost:8080/users/1
 // <User><Id>1</Id><Name>Melissa Raspberry</Name></User>
 //
-func (u *UserService) updateUser(request *restful.Request, response *restful.Response) {
+func (u *UserResource) updateUser(request *restful.Request, response *restful.Response) {
 	usr := new(User)
 	err := request.ReadEntity(&usr)
 	if err == nil {
@@ -82,7 +97,7 @@ func (u *UserService) updateUser(request *restful.Request, response *restful.Res
 // PUT http://localhost:8080/users/1
 // <User><Id>1</Id><Name>Melissa</Name></User>
 //
-func (u *UserService) createUser(request *restful.Request, response *restful.Response) {
+func (u *UserResource) createUser(request *restful.Request, response *restful.Response) {
 	usr := User{Id: request.PathParameter("user-id")}
 	err := request.ReadEntity(&usr)
 	if err == nil {
@@ -96,28 +111,30 @@ func (u *UserService) createUser(request *restful.Request, response *restful.Res
 
 // DELETE http://localhost:8080/users/1
 //
-func (u *UserService) removeUser(request *restful.Request, response *restful.Response) {
+func (u *UserResource) removeUser(request *restful.Request, response *restful.Response) {
 	id := request.PathParameter("user-id")
 	delete(u.users, id)
 }
 
 func main() {
-	u := UserService{map[string]User{}}
-	u.Register()
+	wsContainer := restful.NewContainer()
+	u := UserResource{map[string]User{}}
+	u.Register(wsContainer)
 
 	// Optionally, you can install the Swagger Service which provides a nice Web UI on your REST API
 	// You need to download the Swagger HTML5 assets and change the FilePath location in the config below.
 	// Open http://localhost:8080/apidocs and enter http://localhost:8080/apidocs.json in the api input field.
 	config := swagger.Config{
-		WebServices:    restful.RegisteredWebServices(), // you control what services are visible
+		WebServices:    wsContainer.RegisteredWebServices(), // you control what services are visible
 		WebServicesUrl: "http://localhost:8080",
 		ApiPath:        "/apidocs.json",
 
 		// Optionally, specifiy where the UI is located
 		SwaggerPath:     "/apidocs/",
 		SwaggerFilePath: "/Users/emicklei/Downloads/swagger-ui-1.1.7"}
-	swagger.InstallSwaggerService(config)
+	swagger.RegisterSwaggerService(config, wsContainer)
 
 	log.Printf("start listening on localhost:8080")
-	log.Fatal(http.ListenAndServe(":8080", nil))
+	server := &http.Server{Addr: ":8080", Handler: wsContainer}
+	log.Fatal(server.ListenAndServe())
 }