+++ swagger/swagger_webservice.go
@@ -12,7 +12,14 @@ var config Config
 
 // InstallSwaggerService add the WebService that provides the API documentation of all services
 // conform the Swagger documentation specifcation. (https://github.com/wordnik/swagger-core/wiki).
+// DEPRECATED , use RegisterSwaggerService(...)
 func InstallSwaggerService(aSwaggerConfig Config) {
+	RegisterSwaggerService(aSwaggerConfig, restful.DefaultContainer)
+}
+
+// RegisterSwaggerService add the WebService that provides the API documentation of all services
+// conform the Swagger documentation specifcation. (https://github.com/wordnik/swagger-core/wiki).
+func RegisterSwaggerService(aSwaggerConfig Config, wsContainer *restful.Container) {
 	config = aSwaggerConfig
 
 	ws := new(restful.WebService)
@@ -27,12 +34,12 @@ func InstallSwaggerService(aSwaggerConfig Config) {
 	ws.Route(ws.GET("/{a}/{b}/{c}/{d}/{e}/{f}").To(getDeclarations))
 	ws.Route(ws.GET("/{a}/{b}/{c}/{d}/{e}/{f}/{g}").To(getDeclarations))
 	log.Printf("[restful/swagger] listing is available at %v%v", config.WebServicesUrl, config.ApiPath)
-	restful.Add(ws)
+	wsContainer.Add(ws)
 
 	// Check paths for UI serving
 	if config.SwaggerPath != "" && config.SwaggerFilePath != "" {
 		log.Printf("[restful/swagger] %v%v is mapped to folder %v", config.WebServicesUrl, config.SwaggerPath, config.SwaggerFilePath)
-		http.Handle(config.SwaggerPath, http.StripPrefix(config.SwaggerPath, http.FileServer(http.Dir(config.SwaggerFilePath))))
+		wsContainer.Handle(config.SwaggerPath, http.StripPrefix(config.SwaggerPath, http.FileServer(http.Dir(config.SwaggerFilePath))))
 	} else {
 		log.Printf("[restful/swagger] Swagger(File)Path is empty ; no UI is served")
 	}
@@ -151,7 +158,7 @@ func asModelProperty(sf reflect.StructField, api *Api) ModelProperty {
 		// add|overwrite mode for element type
 		addModelToApi(api, st.Elem())
 	} else {
-		prop.Type = st.String() // inclue pkg path
+		prop.Type = st.String() // include pkg path
 	}
 	return prop
 }
@@ -165,7 +172,7 @@ func asSwaggerParameter(param restful.ParameterData) Parameter {
 		Required:    param.Required}
 }
 
-// Between 1..7 path parameters supported
+// Between 1..7 path parameters is supported
 func composeRootPath(req *restful.Request) string {
 	path := "/" + req.PathParameter("a")
 	b := req.PathParameter("b")
@@ -208,6 +215,8 @@ func asParamType(kind int) string {
 		return "query"
 	case kind == restful.BODY_PARAMETER:
 		return "body"
+	case kind == restful.HEADER_PARAMETER:
+		return "header"
 	}
 	return ""
 }