+++ web_service.go
@@ -37,6 +37,7 @@ func (w *WebService) Param(parameter *Parameter) *WebService {
 }
 
 // PathParameter creates a new Parameter of kind Path for documentation purposes.
+// It is initialized as required with string as its DataType.
 func (w *WebService) PathParameter(name, description string) *Parameter {
 	p := &Parameter{&ParameterData{Name: name, Description: description, Required: true, DataType: "string"}}
 	p.bePath()
@@ -44,6 +45,7 @@ func (w *WebService) PathParameter(name, description string) *Parameter {
 }
 
 // QueryParameter creates a new Parameter of kind Query for documentation purposes.
+// It is initialized as not required with string as its DataType.
 func (w *WebService) QueryParameter(name, description string) *Parameter {
 	p := &Parameter{&ParameterData{Name: name, Description: description, Required: false, DataType: "string"}}
 	p.beQuery()
@@ -51,9 +53,18 @@ func (w *WebService) QueryParameter(name, description string) *Parameter {
 }
 
 // BodyParameter creates a new Parameter of kind Body for documentation purposes.
+// It is initialized as required without a DataType.
 func (w *WebService) BodyParameter(name, description string) *Parameter {
 	p := &Parameter{&ParameterData{Name: name, Description: description, Required: true}}
 	p.beBody()
+	return p
+}
+
+// HeaderParameter creates a new Parameter of kind (Http) Header for documentation purposes.
+// It is initialized as not required with string as its DataType.
+func (w *WebService) HeaderParameter(name, description string) *Parameter {
+	p := &Parameter{&ParameterData{Name: name, Description: description, Required: false, DataType: "string"}}
+	p.beHeader()
 	return p
 }